get expotime from captureduration when cur-expo is null for camera1 dngs

+++ app/src/main/java/freed/cam/apis/camera1/parameters/device/AbstractDevice.java
@@ -233,7 +233,7 @@ public abstract class AbstractDevice implements I_Device {
             return 0;
     }
     @Override
-    public float getCurrentExposuretime()
+    public long getCurrentExposuretime()
     {
         return 0;
     }
